+++ CHANGELOG.md
@@ -4,10 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.5] - ??
+## [1.0.0] - 2022-06-07
+### Added
+- Support for multiprocessing, enabled by default (Issue #46). The number of processes can
+  be controlled through `n_jobs` in `fp.run` and `fp.run_from_iterable`.
+- New interaction: van der Waals contact, based on the sum of vdW radii of two atoms.
+- Saving/loading the fingerprint object as a pickle with `fp.to_pickle` and
+  `Fingerprint.from_pickle` (Issue #40).
+### Changed
+- Molecule suppliers can now be indexed, reused and can return their length, instead of
+  being single-use generators.
 ### Fixed
+- ProLIF can now be installed through pip and conda (Issue #6).
 - If no interaction is detected in the first frame, `to_dataframe` will not complain about
-  a `KeyError` anymore.
+  a `KeyError` anymore (Issue #44).
+- When creating a `plf.Fingerprint`, unknown interactions will no longer fail silently.
 
 ## [0.3.4] - 2021-09-28
 ### Added
@@ -49,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   separated in two lists, one for the ligand and one for the protein (PR #19)
 ### Fixed
 - Residues with a resnumber of `0` are not converted to `None` anymore (Issue #13)
+- Fingerprint instantiated with an unknown interaction name will now raise a `NameError`
 
 ## [0.3.1] - 2021-02-02
 ### Added